+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Renders login button on SSR
 - Loads login content lazily
+- Uses internal, lighter OutsideClickHandler
 
 ## [2.22.0] - 2020-01-09